+++ app/src/components/FilterableSearch/Legend.jsx
